QuestionsController@edit,update,delete用Gateを作成、適用(22)

+++ app/Http/Controllers/QuestionsController.php
@@ -72,6 +72,14 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if (\Gate::allows('update-question', $question)) {
+        //     return view('questions.edit', compact('question'));
+        // }
+        // abort(403, 'Access denied');
+        // 下記と上記は同値、下記はコードの追加のみでおｋなのでdeniesの方がベターかも
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -84,6 +92,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated.');
     }
@@ -96,6 +107,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
     }
